chore: move smsProvider to utils/shared/sms
